Finshed separating functions into different files
