+++ lib/pull.js
@@ -1,5 +1,5 @@
 
-var emp = require('.')
+var gitClone = require('./git-clone')
 var logger = require('./logger')
 var client = require('empirical-client')
 
@@ -9,7 +9,7 @@ exports.pull = function (experiment) {
     return client.getKeys(build.repo.full_name).then(function (keys) {
       console.log('Cloning from', build.repo.ssh_url)
       if (build.push.head_sha) console.log(`Checking out ${build.push.head_sha}`)
-      return emp.getCodeDir(build.repo.ssh_url, build.push.head_sha, keys)
+      return gitClone(build.repo.ssh_url, build.push.head_sha, keys)
     }).then(function (dir) {
       return {
         code_dir: dir,
