+++ lib/run.js
@@ -1,9 +1,13 @@
 
-var emp = require('.')
 var path = require('path')
 var logger = require('./logger')
 var colors = require('colors')
 var initDirs = require('./init-dirs')
+var readProtocol = require('./read-protocol')
+var buildImage = require('./build-image')
+var cache = require('dataset-cache')
+var runExperiment = require('./run-experiment')
+var results = require('./results')
 
 module.exports = function (experiment_name, code_dir) {
   // Get code dir
@@ -12,34 +16,39 @@ module.exports = function (experiment_name, code_dir) {
   code_dir = path.resolve(process.cwd(), code_dir)
   // Get experiment configuration
   logger.section('EXPERIMENT:')
-  var experiment = emp.readExperimentConfig(code_dir, {
-    name: experiment_name
-  })
+  var experiment = readProtocol(code_dir, experiment_name)
+  if (!experiment) return Promise.reject(new Error(`Experiment "${experiment_name}" not found`))
   logger.json(experiment)
   // Build docker Image
   return initDirs().then(function () {
     logger.section('BUILD:')
-    return emp.buildImage(experiment.environment, code_dir, function (data) {
+    return buildImage(experiment.environment, code_dir, function (data) {
       process.stdout.write(data)
     })
   })
   // Get dataset
   .then(function () {
     logger.section('DATASET:')
-    return emp.getDataset(code_dir, experiment.dataset).then(function (data) {
-      if (!data) console.log('No dataset provided')
+    if (!experiment.dataset) {
+      console.log('No dataset provided')
+      return Promise.resolve()
+    }
+    const data = require(path.join(code_dir, experiment.dataset))
+    return cache.install(data, process.env.DATA_PATH)
+    .then(function (data) {
+      console.log('data:', data)
       logger.json(data)
     })
   })
   // Run experiment
   .then(function () {
     logger.section('RUN:')
-    return emp.runExperiment(experiment, logger.write)
+    return runExperiment(experiment, logger.write)
   })
   // Get Results
   .then(function () {
     logger.section('RESULTS:')
-    return emp.getResults(experiment).then(function (overall) {
+    return results.overall(experiment).then(function (overall) {
       logger.json({overall: overall})
       console.log(colors.green.bold('Success'))
     })
