+++ test/index.js
@@ -64,15 +64,80 @@ describe('auth', function () {
   it('.logout() should clear credentials')
 })
 
-describe('data', function () {
-  it('.get(url) should save and log dataset')
-  it('hash file should log the hash of the file')
-})
-
 describe('usage', function () {
   it('.main() should describe the main usage')
   it('.data() should describe the data subcommand usage')
   it('.run() should describe the run subcommand usage')
+})
+
+describe('gitClone', function () {
+  var gitClone = require('../lib/git-clone')
+  it('should clone a repo into a temp directory', function (done) {
+    this.timeout(300000)
+    var repo = 'git@github.com:empiricalci/hello-world.git'
+    var keys = {
+      public_key: fs.readFileSync('./node_modules/fixtures/test_keys/test_key.pub', 'utf8'),
+      private_key: fs.readFileSync('./node_modules/fixtures/test_keys/test_key', 'utf8')
+    }
+    var sha = 'a574f888bdb8f286fd827263794b8aace413dcec'
+    gitClone(repo, sha, keys).then(function (codeDir) {
+      assert(fs.lstatSync(codeDir).isDirectory())
+      done()
+    }).catch(done)
+  })
+  it('should cleanup code and credentials')
+})
+
+describe('readProtocol', function () {
+  const readProtocol = require('../lib/read-protocol')
+  it('should return a valid protocol', function () {
+    var protocol = readProtocol('./node_modules/fixtures/standalone_project', 'hello-world')
+    assert.equal(protocol.type, 'standalone')
+    assert(protocol.environment.tag)
+  })
+  it('should return null if the protocol doesn\'t exits in the empirical.yml', function () {
+    var protocol = readProtocol('./node_modules/fixtures/standalone_project', 'some-protocol')
+    assert(!protocol)
+  })
+})
+
+describe('data', function () {
+  it('.get(url) should save and log dataset')
+  it('hash file should log the hash of the file')
+  it('should install from json file path')
+  it('should install from object')
+})
+
+describe('buildImage', function () {
+  it('should reject if there is an error', function (done) {
+    this.timeout(60000)
+    const buildImage = require('../lib/build-image')
+    buildImage({
+      build: '.',
+      dockerfile: 'bad_dockerfile'
+    }, './test', logHandler).then(function () {
+      done(new Error('Build error not caught'))
+    }).catch(function (err) {
+      assert(err)
+      done()
+    })
+  })
+})
+
+describe('runExperiment', function () {
+  it('should run a sandalone experiment', function (done) {
+    this.timeout(300000)
+    const runExperiment = require('../lib/run-experiment')
+    runExperiment({
+      id: 'some_id',
+      type: 'standalone',
+      environment: {
+        tag: 'empiricalci/test_standalone'
+      }
+    }).then(function () {
+      done()
+    }).catch(done)
+  })
 })
 
 describe('run()', function () {
@@ -93,64 +158,14 @@ describe('run()', function () {
       done()
     })
   })
-  it('should fail if the experiment-name is not found')
-})
-
-describe('Library', function () {
-  var emp = require('../lib')
-  it('should clone a repo into a temp directory', function (done) {
-    this.timeout(300000)
-    var repo = 'git@github.com:empiricalci/hello-world.git'
-    var keys = {
-      public_key: fs.readFileSync('./node_modules/fixtures/test_keys/test_key.pub', 'utf8'),
-      private_key: fs.readFileSync('./node_modules/fixtures/test_keys/test_key', 'utf8')
-    }
-    var sha = 'a574f888bdb8f286fd827263794b8aace413dcec'
-    emp.getCodeDir(repo, sha, keys).then(function (codeDir) {
-      assert(fs.lstatSync(codeDir).isDirectory())
+  it('should fail if the experiment-name is not found', function (done) {
+    run('something', 'node_modules/fixtures/standalone_project')
+    .then(function () {
+      done(new Error('Experiment not found error wasn\'t caught'))
+    })
+    .catch(function (err) {
+      assert.equal(err.message, `Experiment "something" not found`)
       done()
-    }).catch(done)
-  })
-  describe('readExperimentConfig', function () {
-    it('should succed with valid standalone config', function () {
-      var experiment = emp.readExperimentConfig('./node_modules/fixtures/standalone_project', {
-        _id: '342434234',
-        name: 'hello-world',
-        type: 'standalone'
-      })
-      assert.equal(experiment.type, 'standalone')
-      assert(experiment.environment.tag)
-    })
-    it('should fail if a solver experiment config does not contain evaluator')
-  })
-  it('should get a datset')
-  describe('runExperiment', function () {
-    it('should run a sandalone experiment', function (done) {
-      this.timeout(300000)
-      emp.runExperiment({
-        _id: 'some_id',
-        type: 'standalone',
-        environment: {
-          tag: 'empiricalci/test_standalone'
-        }
-      }).then(function () {
-        done()
-      }).catch(done)
-    })
-  })
-  it('should cleanup code and credentials')
-  describe('buildImage', function () {
-    it('should reject if there is an error', function (done) {
-      this.timeout(60000)
-      emp.buildImage({
-        build: '.',
-        dockerfile: 'bad_dockerfile'
-      }, './test', logHandler).then(function () {
-        done(new Error('Build error not caught'))
-      }).catch(function (err) {
-        assert(err)
-        done()
-      })
     })
   })
 })
